Added find_iterator to cifdoc Table
Added find_iterator to cifdoc Table. This function is an alternative to
the find function, which uses exception to signal that nothing was
found.

The new find_iterator function returns an iterator instead of the Row
itself. This is used to return the end() iterator when no corresponding
Row is found instead of throwing exception as in the find function.

This change enables us to use mmCIF files on the WebAssembly platform,
which does not yet support exceptions efficiently enough.

Also changed the find implementation to internally use the find_iterator
function.

Fixed the mmcif loading from using exceptions to using the
find_row_iterator function.